+++ asynctnt/__init__.py
@@ -3,4 +3,4 @@ from .iproto.protocol import (
     Iterator, Response, TarantoolTuple, PushIterator
 )
 
-__version__ = '1.0b2'
+__version__ = '1.0b3'